refactor(activations): Move VarsFromEnvironment
Now that it's used by `cli::activate` and `cli::executive`.
