+++ cli/flox-activations/src/vars_from_env.rs
@@ -1,0 +1,30 @@
+use anyhow::{Result, anyhow};
+use serde::{Deserialize, Serialize};
+
+use crate::activate_script_builder::FLOX_ENV_DIRS_VAR;
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct VarsFromEnvironment {
+    pub flox_env_dirs: Option<String>,
+    pub path: String,
+    pub manpath: Option<String>,
+}
+
+impl VarsFromEnvironment {
+    pub fn get() -> Result<Self> {
+        let flox_env_dirs = std::env::var(FLOX_ENV_DIRS_VAR).ok();
+        let path = match std::env::var("PATH") {
+            Ok(path) => path,
+            Err(e) => {
+                return Err(anyhow!("failed to get PATH from environment: {}", e));
+            },
+        };
+        let manpath = std::env::var("MANPATH").ok();
+
+        Ok(Self {
+            flox_env_dirs,
+            path,
+            manpath,
+        })
+    }
+}
